finish
At first, we did some EDA for unevenly space time series data. We pointed out that we can descirbe some of the data in time behaviours by global situations such as Covid-19. After that, timesteps distributions was calculated, we emphasized that its variation might be little enough to try on imputing method. On the other hand, one of the project goals was to use a neural network, which doesn't require such a transformations. We have decided to use RNNs gated version architecture, GRU, as a first model. We found a hints in a literature about hyperparameter tuning, which helps as to se the first hyperparameter space. Despite the fact that a model from first tuning has some hyperparameters on the borders, second tuning, with a potentially better parameter space, didn't improve that. Early stopping and tensorboard callbacks was used. Thanks to them, we didn't need to specify an ideal epoch count, additionally we were able to visualize a whole training process on the plots and tables. In the last stage, we have evaluated a model on the remaining Amazons data, after which we checked, how well the model generalize for other comapnies under consideration. We found out that the results for Amazon was the worst.